refactor(events): extract MucInvitationItem and StrangerMessageItem
